moved population density calculations into their own folder
